// src/api/models/SearchApartmentsResponseV1ResponseBody.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SITE API
 * API сайта ПИК Брокер
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime'
import { Apartment, ApartmentFromJSON, ApartmentFromJSONTyped, ApartmentToJSON } from './'

/**
 *
 * @export
 * @interface SearchApartmentsResponseV1ResponseBody
 */
export interface SearchApartmentsResponseV1ResponseBody {
  /**
   *
   * @type {number}
   * @memberof SearchApartmentsResponseV1ResponseBody
   */
  count?: number
  /**
   *
   * @type {number}
   * @memberof SearchApartmentsResponseV1ResponseBody
   */
  countPages: number
  /**
   *
   * @type {Array<Apartment>}
   * @memberof SearchApartmentsResponseV1ResponseBody
   */
  apartments: Array<Apartment>
}

export function SearchApartmentsResponseV1ResponseBodyFromJSON(
  json: any,
): SearchApartmentsResponseV1ResponseBody {
  return SearchApartmentsResponseV1ResponseBodyFromJSONTyped(json, false)
}

export function SearchApartmentsResponseV1ResponseBodyFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): SearchApartmentsResponseV1ResponseBody {
  if (json === undefined || json === null) {
    return json
  }
  return {
    count: !exists(json, 'count') ? undefined : json['count'],
    countPages: json['countPages'],
    apartments: (json['apartments'] as Array<any>).map(ApartmentFromJSON),
  }
}

export function SearchApartmentsResponseV1ResponseBodyToJSON(
  value?: SearchApartmentsResponseV1ResponseBody | null,
): any {
  if (value === undefined) {
    return undefined
  }
  if (value === null) {
    return null
  }
  return {
    count: value.count,
    countPages: value.countPages,
    apartments: (value.apartments as Array<any>).map(ApartmentToJSON),
  }
}