// src/api/models/Widget.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SITE API
 * API сайта ПИК Брокер
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime'
import {
  WidgetFields,
  WidgetFieldsFromJSON,
  WidgetFieldsFromJSONTyped,
  WidgetFieldsToJSON,
  WidgetRestrict,
  WidgetRestrictFromJSON,
  WidgetRestrictFromJSONTyped,
  WidgetRestrictToJSON,
} from './'

/**
 *
 * @export
 * @interface Widget
 */
export interface Widget {
  /**
   *
   * @type {string}
   * @memberof Widget
   */
  name: string
  /**
   *
   * @type {{ [key: string]: WidgetFields; }}
   * @memberof Widget
   */
  fields: { [key: string]: WidgetFields }
  /**
   *
   * @type {object}
   * @memberof Widget
   */
  data?: object | null
  /**
   *
   * @type {WidgetRestrict}
   * @memberof Widget
   */
  restrict?: WidgetRestrict | null
}

export function WidgetFromJSON(json: any): Widget {
  return WidgetFromJSONTyped(json, false)
}

export function WidgetFromJSONTyped(json: any, ignoreDiscriminator: boolean): Widget {
  if (json === undefined || json === null) {
    return json
  }
  return {
    name: json['name'],
    fields: mapValues(json['fields'], WidgetFieldsFromJSON),
    data: !exists(json, 'data') ? undefined : json['data'],
    restrict: !exists(json, 'restrict') ? undefined : WidgetRestrictFromJSON(json['restrict']),
  }
}

export function WidgetToJSON(value?: Widget | null): any {
  if (value === undefined) {
    return undefined
  }
  if (value === null) {
    return null
  }
  return {
    name: value.name,
    fields: mapValues(value.fields, WidgetFieldsToJSON),
    data: value.data,
    restrict: WidgetRestrictToJSON(value.restrict),
  }
}