// src/api/models/Promo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SITE API
 * API сайта ПИК Брокер
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Промо акция для расчёта скидки и ипотеки
 * @export
 * @enum {string}
 */
export enum Promo {
  Default = 'default',
  ExMax = 'ex_max',
}

export function PromoFromJSON(json: any): Promo {
  return PromoFromJSONTyped(json, false)
}

export function PromoFromJSONTyped(json: any, ignoreDiscriminator: boolean): Promo {
  return json as Promo
}

export function PromoToJSON(value?: Promo | null): any {
  return value as any
}